implementation/requirements.txt
boto3==1.42.54
botocore==1.42.54
cloudpickle==3.1.2
copulas==0.14.1
ctgan==0.12.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.5
deepecho==0.8.1
Faker==40.4.0
filelock==3.24.3
fsspec==2026.2.0
graphviz==0.21
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.17.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
pandas==2.3.3
platformdirs==4.9.2
plotly==6.5.2
python-dateutil==2.9.0.post0
python-snap7==2.0.2
pytz==2025.2
PyYAML==6.0.3
rdt==1.20.0
s3transfer==0.16.0
scikit-learn==1.6.1
scipy==1.17.1
sdmetrics==0.28.0
sdv==1.35.0
setuptools==82.0.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3